Fix logging calls
When logging, %-style substitution should be used with parameter values
passed to the logging call, not immediately interpolated. There is some
dispute about whether the efficiency gained by avoiding interpolation
when a message is not logged at all outweighs the effiency lost compared
to using more efficient interpolation mechanisms; however, what is not
disputed is that logging frameworks that can be make semantic use of the
parameters require that they be kept separate from the rest of the
message.